compiler fixes
Very rarely the "remove parens" flag to gofmt causes a failure, so I stopped using it (in -d mode of d8mc). Fixed various crashes and bugs in obscure corners of the typechecker. Same for the backend, especially in the area of generating zero values. In module files, the export statement wasn't checked for dups and this caused obscure but serious issues. Similarly, the export statement wasn't checking that each exported symbol was in fact defined.